requirements.txt
about-time==4.2.1
alive-progress==3.2.0
amqp==5.3.1
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.10.0
astroid==3.3.11
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.2
black==25.9.0
bleach==6.2.0
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.4.0
Django==5.2.7
django-bootstrap3==25.2
django-debug-toolbar==6.0.0
django-filter==25.2
django-tables2==2.7.5
duckdb==1.1.2
executing==2.1.0
fastjsonschema==2.20.0
flake8==7.3.0
fonttools==4.54.1
fqdn==1.5.1
grapheme==0.6.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==6.1.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kombu==5.5.4
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
mypy_extensions==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.3
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.6.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.18.0
pylast==5.3.0
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytokens==0.2.0
pytz==2024.2
pywin32==308; sys_platform == "win32"
pywinpty==2.0.14; sys_platform == "win32"
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
Send2Trash==1.8.3
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
spotipy==2.24.0
sqlparse==0.5.3
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
tzdata==2025.2
uri-template==1.3.0
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
